[spark] Introduce a base class for SparkTable to support multi-version APIs
